Release 2.1.3

+++ installSDK.sh
@@ -1,22 +1,35 @@
 #!/usr/bin/bash
 # installSDK.sh
-# CloudBackend 2022-2023.
-export RELEASE="2.1.2"
+# CloudBackend 2022-2024.
+export RELEASE="2.1.3"
 echo "Installing CloudBackend SDK version" ${RELEASE}
 echo    "Minimum   o/s: Ubuntu 20.04"
 echo -n "Installed o/s: "
 cat /etc/issue | cut -d '\' -f1
 mkdir -p ~/cbe                              # if not already created
 mydir=${PWD}
+echo ${mydir}
+if [ "${mydir}" = "$HOME/Downloads" ];
+then
+  echo "Error: May not run from" ${mydir}
+  exit 2
+fi
 echo "moving" ${mydir} "to ~/cbe"
+read -p "continue (Y/n) ? " Answer
+if [ "${Answer}" = "n" ] || [ "${Answer}" = "N" ];
+then
+  echo "exit"
+  exit 1
+fi
 if [ -d ~/cbe/${RELEASE} ];
 then
   ls -ld ~/cbe/${RELEASE} ;
   echo "renaming" ~/cbe/${RELEASE} "to" ~/cbe/removed_${RELEASE}_`date --iso`;
-  mv ~/cbe/${RELEASE} ~/cbe/removed_${RELEASE}_`date --iso`;
-  ls -ld ~/cbe/2*
+  mv ~/cbe/${RELEASE} ~/cbe/removed_${RELEASE}_`date +%FT%H%M`;
+  ls -ld ~/cbe/2* ~/cbe/removed_*
 fi
 mv ${mydir} ~/cbe/${RELEASE}
+touch ~/cbe/${RELEASE}
 cd ~/cbe
 ls -l current
 rm -i current                               # remove old symbolic link
@@ -34,28 +47,21 @@ bash installC++SDK.sh                       # install script
 echo "done."
 
 echo ¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤
-read -p "Next Java. Press ENTER " Answer
-cd ~/cbe/current                            # the release directory
-cd Java                                     # SDK Java
-pwd
-bash installJavaSDK.sh                      # install script
-echo "done."
+read -p "Install SDK Java module (Y/n) " Answer
+if [ "${Answer}" = "n" ] || [ "${Answer}" = "N" ];
+then
+  echo "skip Java";
+else
+  cd ~/cbe/current                            # the release directory
+  cd Java                                     # SDK Java
+  pwd
+  bash installJavaSDK.sh                      # install script
+  echo "done."
+fi
 
 echo ¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤¤
-read -p "Next CLI. Press ENTER " Answer
-echo "Installing CloudBackend CLI."
-cd ~/cbe/current                            # the release directory
-cd tools                                    # cli
-pwd
-unzip new_cli.zip                           # unpack the new_cli
-./new_cli -uc gh1  << EOF                   # start new_cli with user credentials of githubtester1
- help
- set
- logout
-EOF
-echo "done."
 cd
-ls -l cbe/current
-cd cbe/current
-pwd
-
+ls -l ~/cbe/current
+echo "next: go to ~/cbe/current"
+echo
+echo "do: cd .."
